+++ packages/tnix-lsp/src/Server.spec.hs
@@ -39,6 +39,7 @@ spec = do
                 , "documentSymbolProvider" .= True
                 , "workspaceSymbolProvider" .= True
                 , "codeActionProvider" .= True
+                , "inlayHintProvider" .= object ["resolveProvider" .= False]
                 , "semanticTokensProvider"
                     .= object
                       [ "legend"